requirements.txt
fastapi
uvicorn
sqlalchemy
pandas
python-multipart
twilio
python-dotenv